Improved Performance
Used images of WebP format instead of PNG.
The hero section had a different image depending on the screen size
but they were in PNG format so I converted them to WebP. This reduced
their size to almost half and will improve performance considerably.

Used React.lazy() and Suspense to lazy load JS. Also specified a
min-height for the hero-mobile image to reserve space for it in
order to reduce CLS. This was recommended by PageSpeed Insights.